requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
pydantic==2.3.0
pymongo==4.5.0
folium==0.14.0
python-dotenv
mongoengine